requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
jinja2==3.1.3
python-multipart==0.0.9
firebase-admin==6.4.0
requests==2.31.0
beautifulsoup4==4.12.3
fuzzywuzzy==0.18.0
python-Levenshtein==0.25.0
SQLAlchemy==2.0.27
psycopg2-binary==2.9.9
python-dotenv==1.0.1 